requirements.txt
chromadb>=0.4.0
langchain>=0.1.0
langchain-chroma>=0.1.0
langchain-community>=0.0.0
langchain-ollama>=0.0.0
streamlit>=1.30.0
unstructured>=0.12.0
tika>=2.6.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
rich>=13.0.0
httpx>=0.26.0
pandas>=2.0.0
fastapi>=0.109.0
uvicorn>=0.27.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0